Consolidate Sentry policy behind runtime adapters
Keep the deployed root/Deno reporter compatible while adding explicit Node and Deno adapter entrypoints for downstream service migration. Shared policy owns redaction, tags, fingerprinting, Grafana trace correlation, and failure-isolated capture/flush behavior.

Constraint: Stack on agent/sentry-v2-extension-entrypoints and keep Sentry error-only without OpenTelemetry or tracing ownership.

Rejected: Reusing the root export for Node consumers | it would keep Node services coupled to the Deno SDK entrypoint.

Confidence: high

Scope-risk: moderate

Directive: Keep service-local Sentry logic out of TypeScript services; use explicit ./node or ./deno adapters instead.

Tested: deno test --allow-all extensions/ext-observability-sentry/src/policy.test.ts extensions/ext-observability-sentry/src/deno.test.ts extensions/ext-observability-sentry/src/node.test.ts extensions/ext-observability-sentry/src/index.test.ts; deno test --config=scripts/test.deno.json --no-check --allow-read --allow-write --allow-run scripts/build/npm-extension-package-metadata.test.ts scripts/build/npm-package-metadata.test.ts; deno task build:npm; deno task typecheck:consumer; deno task typecheck; generated package import/isolation smoke checks

Not-tested: deno task verify:quick is blocked by unrelated base-branch docs drift for veryfront/release-assets missing @example and docs/api-reference/veryfront/release-assets.md.
