+++ extensions/ext-observability-sentry/README.md
@@ -14,3 +14,15 @@ binaries include the dormant adapter; npm consumers install
 `@veryfront/ext-observability-sentry` separately. The adapter captures
 unexpected application failures, tags them by service and boundary, and keeps
 OpenTelemetry as the owner of traces, metrics, and logs.
+
+## Runtime entrypoints
+
+- `@veryfront/ext-observability-sentry` keeps the V1 Deno-compatible root
+  reporter export.
+- `@veryfront/ext-observability-sentry/deno` exports the explicit Deno reporter.
+- `@veryfront/ext-observability-sentry/node` exports the explicit Node reporter.
+
+The Node and Deno reporters share the same privacy policy, service tags,
+`veryfront.boundary` tagging, Grafana trace correlation, fingerprinting, and
+bounded flush behavior. They only use Sentry for error capture; tracing, logs,
+request bodies, user data, and OpenTelemetry provider setup remain disabled.
