+++ extensions/ext-observability-sentry/deno.json
@@ -1,7 +1,11 @@
 {
   "name": "@veryfront/ext-observability-sentry",
   "version": "0.1.0",
-  "exports": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./deno": "./src/deno.ts",
+    "./node": "./src/node.ts"
+  },
   "veryfront": {
     "extension": true,
     "capabilities": [
@@ -10,6 +14,7 @@
   },
   "imports": {
     "@sentry/deno": "npm:@sentry/deno@10.68.0",
+    "@sentry/node": "npm:@sentry/node@10.68.0",
     "@std/assert": "jsr:@std/assert@1.0.19",
     "veryfront/extensions": "../../src/extensions/index.ts",
     "veryfront/observability/sentry": "../../src/observability/sentry.ts"
