+++ extensions/ext-observability-sentry/src/policy.ts
@@ -1,0 +1,145 @@
+import type { ApplicationErrorContext } from "veryfront/observability/sentry";
+
+export const DEFAULT_FINGERPRINT = "{{ default }}";
+
+const SENTRY_TOKEN_PATTERN = /\bsntrys_[A-Za-z0-9_+/=-]+\b/g;
+const BEARER_TOKEN_PATTERN = /\bBearer\s+[A-Za-z0-9._~+/=-]+\b/gi;
+const JWT_PATTERN = /\b[A-Za-z0-9_-]{16,}\.[A-Za-z0-9_-]{16,}\.[A-Za-z0-9_-]{16,}\b/g;
+const URL_CREDENTIAL_PATTERN = /(https?:\/\/)[^@\s/]+(@[^/\s]+)/gi;
+const SENSITIVE_QUERY_PATTERN =
+  /([?&](?:access_token|api_key|auth|authorization|dsn|password|secret|token)=)[^&#\s]+/gi;
+
+export type SentryPolicyScope = {
+  setContext(name: string, context: Record<string, unknown>): void;
+  setFingerprint(fingerprint: string[]): void;
+  setTag(key: string, value: string): void;
+};
+
+export type SentryPolicySdk = {
+  captureException(error: unknown): string;
+  flush(timeoutMs?: number): Promise<boolean>;
+  withScope(callback: (scope: SentryPolicyScope) => void): void;
+};
+
+export type SentryPolicyEvent = {
+  breadcrumbs?: unknown;
+  exception?: {
+    values?: Array<{
+      value?: string;
+      stacktrace?: {
+        frames?: Array<{
+          abs_path?: string;
+          filename?: string;
+        }>;
+      };
+    }>;
+  };
+  fingerprint?: string[];
+  logentry?: {
+    message?: string;
+  };
+  message?: string;
+  request?: unknown;
+  tags?: Record<string, unknown>;
+  type?: undefined;
+  user?: unknown;
+};
+
+export function captureWithSentryPolicy(
+  sdk: SentryPolicySdk,
+  serviceName: string,
+  error: unknown,
+  context: ApplicationErrorContext,
+): string | undefined {
+  try {
+    let eventId: string | undefined;
+    sdk.withScope((scope) => {
+      applySentryScopePolicy(scope, serviceName, context);
+      eventId = sdk.captureException(error);
+    });
+    return eventId;
+  } catch {
+    return undefined;
+  }
+}
+
+export async function flushWithSentryPolicy(
+  sdk: Pick<SentryPolicySdk, "flush">,
+  timeoutMs?: number,
+): Promise<boolean> {
+  try {
+    return await sdk.flush(timeoutMs);
+  } catch {
+    return false;
+  }
+}
+
+export function applySentryScopePolicy(
+  scope: SentryPolicyScope,
+  serviceName: string,
+  context: ApplicationErrorContext,
+): void {
+  scope.setFingerprint([serviceName, DEFAULT_FINGERPRINT]);
+  scope.setTag("service.name", serviceName);
+  scope.setTag("veryfront.boundary", context.boundary);
+  if (context.method) scope.setTag("http.request.method", context.method);
+  if (context.requestId) scope.setTag("veryfront.request_id", context.requestId);
+  if (context.traceId) {
+    scope.setTag("grafana.trace_id", context.traceId);
+    scope.setContext("grafana_trace", {
+      trace_id: context.traceId,
+      ...(context.spanId ? { span_id: context.spanId } : {}),
+    });
+  }
+}
+
+export function prepareSentryEvent<TEvent extends SentryPolicyEvent>(
+  event: TEvent,
+  serviceName: string,
+): TEvent & SentryPolicyEvent {
+  event.tags = {
+    ...event.tags,
+    "service.name": serviceName,
+  };
+  event.fingerprint = [serviceName, ...(event.fingerprint ?? [DEFAULT_FINGERPRINT])];
+
+  delete event.breadcrumbs;
+  delete event.request;
+  delete event.user;
+
+  if (event.message) event.message = redactSensitiveText(event.message);
+  if (event.logentry?.message) {
+    event.logentry.message = redactSensitiveText(event.logentry.message);
+  }
+  for (const value of event.exception?.values ?? []) {
+    if (value.value) value.value = redactSensitiveText(value.value);
+    for (const frame of value.stacktrace?.frames ?? []) {
+      if (frame.filename) frame.filename = sanitizeStackFramePath(frame.filename);
+      if (frame.abs_path) frame.abs_path = sanitizeStackFramePath(frame.abs_path);
+    }
+  }
+
+  return event;
+}
+
+export function redactSensitiveText(value: string): string {
+  return value
+    .replace(SENTRY_TOKEN_PATTERN, "[REDACTED]")
+    .replace(BEARER_TOKEN_PATTERN, "[REDACTED]")
+    .replace(JWT_PATTERN, "[REDACTED]")
+    .replace(URL_CREDENTIAL_PATTERN, "$1[REDACTED]$2")
+    .replace(SENSITIVE_QUERY_PATTERN, "$1[REDACTED]");
+}
+
+export function sanitizeStackFramePath(value: string): string {
+  const redacted = redactSensitiveText(value);
+  const path = redacted.startsWith("file://") ? redacted.slice("file://".length) : redacted;
+  if (!path.startsWith("/") && !/^[A-Za-z]:[\\/]/.test(path)) return redacted;
+
+  const normalized = path.replaceAll("\\", "/");
+  const sourcePath = normalized.match(/(?:^|\/)((?:src|extensions|cli|dist)\/.+)$/)?.[1];
+  if (sourcePath) return sourcePath;
+
+  const basename = normalized.split("/").filter(Boolean).at(-1);
+  return basename ? `[REDACTED]/${basename}` : "[REDACTED]";
+}
